304: Move around the CAN models to be in the CANS package
